No longer attempting to create a user record from commits, pulls, and issues.
Instead just saving off username with repo activity and linking their GitHub profile. The ultimate goal of this website is to place a layer over GitHub not make a social profile of a user's activity. In the future we can find a user's activity by searching tables for their username if we wanted to.